Unify parsing.
