+++ query-engine/core/src/query_graph_builder/read/aggregations/mod.rs
@@ -1,39 +1,14 @@
 use super::*;
-use crate::{query_document::ParsedField, AggregateRecordsQuery, FieldPair, ReadQuery};
+
+mod aggregate;
+mod group_by;
+
+pub use aggregate::*;
+pub use group_by::*;
+
+use crate::FieldPair;
 use connector::AggregationSelection;
 use prisma_models::{ModelRef, ScalarFieldRef};
-
-pub fn aggregate(field: ParsedField, model: ModelRef) -> QueryGraphBuilderResult<ReadQuery> {
-    let name = field.name;
-    let alias = field.alias;
-    let model = model;
-    let nested_fields = field.nested_fields.unwrap().fields;
-    let selection_order = collect_selection_tree(&nested_fields);
-    let args = extractors::extract_query_args(field.arguments, &model)?;
-
-    // Reject unstable cursors for aggregations, because we can't do post-processing on those (we haven't implemented a in-memory aggregator yet).
-    if args.contains_unstable_cursor() {
-        return Err(QueryGraphBuilderError::InputError(
-            "The chosen cursor and orderBy combination is not stable (unique) and can't be used for aggregations."
-                .to_owned(),
-        ));
-    }
-
-    let selectors: Vec<_> = nested_fields
-        .into_iter()
-        .map(|field| resolve_query(field, &model))
-        .collect::<QueryGraphBuilderResult<_>>()?;
-
-    Ok(ReadQuery::AggregateRecordsQuery(AggregateRecordsQuery {
-        name,
-        alias,
-        model,
-        selection_order,
-        args,
-        selectors,
-        group_by: vec![],
-    }))
-}
 
 /// Resolves the given field as a aggregation query.
 fn resolve_query(field: FieldPair, model: &ModelRef) -> QueryGraphBuilderResult<AggregationSelection> {
@@ -43,7 +18,7 @@ fn resolve_query(field: FieldPair, model: &ModelRef) -> QueryGraphBuilderResult<
         "sum" => AggregationSelection::Sum(resolve_fields(model, field)),
         "min" => AggregationSelection::Min(resolve_fields(model, field)),
         "max" => AggregationSelection::Max(resolve_fields(model, field)),
-        _ => unreachable!(),
+        name => AggregationSelection::Field(model.fields().find_from_scalar(name).unwrap()),
     };
 
     Ok(query)
